+++ cqi_cpp/include/qtree.hpp
@@ -2,15 +2,15 @@
 
 class QTree: public QFunc {
     public:
-        float splitThreshMax, splitThreshDecay, splitThresh; 
+        double splitThreshMax, splitThreshDecay, splitThresh; 
         QTreeNode* root;
         bool _justSplit;
-        unordered_map<string, float>* params;
+        unordered_map<string, double>* params;
 
-        QTree(Box*, Discrete*, QTreeNode*, float, float, float, float, float, int);
+        QTree(Box*, Discrete*, QTreeNode*, double, double, double, double, double, int);
         int selectA(State*);
-        void takeTuple(State*, Action*, float, State*, bool);
-        void update(State*, Action*, float, State*, bool);
+        void takeTuple(State*, Action*, double, State*, bool);
+        void update(State*, Action*, double, State*, bool);
         int numNodes();
         void printStructure();
         bool justSplit();
